requirements.txt
mcp[cli]>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0